Upload route diagrams for M31 extension
Upload the platform and in-carriage route diagrams for the M31 extension. M31 previously did not have either type of route diagram, but the extension of the line necessitates the creation of these diagrams. Implemented changes in the platform route diagrams compared to the diagrams for other lines. These were suggested by Janukiii_, and see the removal of the HarlonTransit footer to allow for more space and moving mode icons to be below station names.